requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.16.0
dnspython==2.8.0
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.27.0